Potential fix for pull request finding

+++ secrets.example.h
@@ -6,4 +6,4 @@
 constexpr const char* WIFI_SSID = "CHANGE_ME_SSID";
 constexpr const char* WIFI_PASSWORD = "CHANGE_ME_PASSWORD";
 constexpr const char* MQTT_SERVER_IP = "127.0.0.1";
-constexpr unsigned int MQTT_SERVER_PORT = 1883;
+constexpr uint16_t MQTT_SERVER_PORT = 1883;
